Move the folders movies, top, and tv outside to collection folder
