+++ src/core/main/use-initialice-movies.js
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useDispatch } from "react-redux";
 //own
-import { getMovies } from "store/collections/movies/get-movies";
+import { getMovies } from "store/movies/get-movies";
 
 export default function useInitialiceMovies() {
   const dispatch = useDispatch();
